data_cholesterol {grafify}	R Documentation

Hierarchical data from 25 subjects either treated or not at 5 hospitals - two-way ANOVA design with repeated measures.

Description

An example dataset on measurements of blood cholesterol levels measured in 5 subjects measured before and after receiving a Drug. Five patients each were recruited at 5 hospitals (a-e), so that there are 25 different subjects (1-25) measured twice. Data are from Micro/Immuno Stats

Usage

data_cholesterol

Format

tibble: 30 obs. of 3 variables:

Hospital

Factor with 5 levels (a-e), representing different hospitals where subjects were recruited.

Subject

A factor with 25 levels denoting individuals on whom measurements were made twice.

Treatment

A factor with 2 levels indicating when measurements were made, i.e. before and after drug.

Cholesterol

Numerical dependent variable indicating measured doubling time in min.

[Package grafify version 5.1.0 Index]