data_doubling_time {grafify}	R Documentation

Doubling time of E.coli measured by 10 students three independent times.

Description

An example dataset showing measurements of E. coli doubling times (in min) measured by 10 different students in 3 independent experiments each. Note that Experiments are just called Exp1-Exp3 even though Exp1 of any of the students are not connected in anyway - this will confuse R! Data are from Micro/Immuno Stats

Usage

data_doubling_time

Format

tibble: 30 obs. of 3 variables:

Student

Factor with 10 levels, representing different students.

Experiment

A factor with 3 levels representing independent experiments.

Doubling_time

Numerical dependent variable indicating measured doubling time in min.

[Package grafify version 5.1.0 Index]