data_t_pdiff {grafify}	R Documentation

Matched data from two groups where difference between them is consistent.

Description

An example dataset for paired difference Student's t test. These are bodyweight (Mass) in grams of same mice left untreated or treated, which are two groups to be compared. The data are in a longtable format, and the two groups are levels within the factor "Condition". The Subject column lists ID of matched mice that were measured without and with treatment. These data are from Sanchez-Garrido et al, Sci Signal, 2018 (doi: 10.1126/scisignal.aat6903).

Usage

data_t_pdiff

Format

data.frame: 20 obs. of 3 variables:

Subject

Factor with 10 levels, denoted by capital letters, representing individuals or subjects.

Condition

A fixed factor with 2 levels: "Untreated" & "Treated".

Mass

Numerical dependent variable indicating body mass of mice

[Package grafify version 5.1.0 Index]