data_t_pratio {grafify}	R Documentation

Matched data from two groups where ratio between them is consistent.

Description

An example dataset for paired ratio Student's t test. These are Cytokine measurements by ELISA (in ng/ml) from 33 independent in vitro experiments performed on two Genotypes that we want to compare. The data are in a longtable format, and the two groups are levels within the factor "Genotype". The Experiment column lists ID of matched experiments.

Usage

data_t_pratio

Format

data.frame: 66 obs. of 3 variables:

Genotype

Factor with 2 levels, representing genotypes to be compared ("WT" & "KO").

Experiment

A random factor with 33 levels representing independent experiments, denoted as "Exp_1", "Exp_2"...

Cytokine

Numerical dependent variable indicating cytokine measured by ELISA.

[Package grafify version 5.1.0 Index]