data_zooplankton {grafify}	R Documentation

Time-series data on zooplankton in lake Menon.

Description

A subset of data from (Lathro RC, 2000) (doi: 10.6073/pasta/ec3d0186753985147d4f283252388e05) provided by the Wisconsin Department of Natural Resources

Usage

data_zooplankton

Format

tibble: 1127 obs. of 8 variables:

day

Numeric integer variable.

year

Numeric integer variable of years during which data were collected.

lake

This data is for lake Menon; data for other others not included in this subset.

taxon

Names of zooplankton taxa as factor of 8 levels.

density

Numeric values of density of measurements.

density_adj

Numeric values of adjusted density .

min_density

Numeric values of minimum densities.

desnsity_scaled

Numeric value of scaled density.

[Package grafify version 5.1.0 Index]