graf_colours {grafify}	R Documentation

List of hexcodes of colours in grafify palettes

Description

To visualise these colours use plot_grafify_palette. okabe_ito, bright, contrast, dark, light, muted, pale, vibrant, yello_conti from Paul Tol's colours (also see khroma package). Zesty, Pastel, Elegant from this link. Colour hexcodes for fishy, kelly, r4, safe, OrBl_div, PrGn_div, blue_conti, grey_conti taken from cols4all:c4a_gui package. All schemes are colour blind-friendly.

Usage

graf_colours

Format

An object of class character of length 154.

Value

This is a character vector with names and hexcodes of colours used by palette functions. It is used by get_graf_colours to generate palettes.

[Package grafify version 5.1.0 Index]