graf_palettes {grafify}	R Documentation

List of palettes available in grafify package

Description

To visualise these colours use plot_grafify_palette.

Usage

graf_palettes

Format

An object of class list of length 18.

Value

This function returns a list of palettes in grafify with names and hexcodes of colours in those palettes. Names of palettes available are as follows:

Categorical/discreet palettes:

okabe_ito

bright

contrast

dark

kelly

light

muted

pale

r4

safe

vibrant

Sequential quantitative palettes:

grey_conti

blue_conti

yellow_conti

Divergent quantitative palettes:

OrBl_div

PrGn_div

[Package grafify version 5.1.0 Index]