simple_anova {grafify}	R Documentation

ANOVA table from a linear model fit to data.

Description

One of two functions for simple ANOVA tables and linear models without random effects, which use lm to fit a linear models.

link{simple_anova}

link{simple_model}

Usage

simple_anova(data, Y_value, Fixed_Factor, ...)

Arguments

data
a data table object, e.g. data.frame or tibble.

Y_value
name of column containing quantitative (dependent) variable, provided within "quotes". The following transformations are permitted: "log(Y_value)", "log(Y_value + c)" where c a positive number, "logit(Y_value)" or "logit(Y_value/100)" which may be useful when Y_value are percentages (note quotes outside the log or logit calls); "sqrt(Y_value)" or "(Y_value)^2" should also work. During posthoc-comparisons, log and logit transformations will be back-transformed to the original scale. Other transformations, e.g., "sqrt(Y_value)" will not be back-transformed. Check out the regrid and ref_grid for details if you need back-transformation to the response scale.

Fixed_Factor
name(s) of categorical fixed factors (independent variables) provided within quotes (e.g., "A") or as a vector if more than one (e.g., c("A", "B"). If a numeric variable(s) is used, transformations similar to Y_value are permitted.

...
any additional argument to pass on to lm if required.

Details

Update in v0.2.1: This function uses lm to fit a linear model to data, passes it on to Anova, and outputs the ANOVA table with type II sum of squares with F statistics and P values. (Previous versions produced type I sum of squares using anova call.)

It requires a data table, one quantitative dependent variable and one or more independent variables. If your experiment design has random factors, use the related function mixed_anova.

This function is related to link{simple_model}.

Value

ANOVA table of class "anova" and "data.frame".

Examples

#Basic usage 
simple_anova(data = data_doubling_time, 
Y_value = "Doubling_time", 
Fixed_Factor = "Student")

[Package grafify version 5.1.0 Index]