Model from a linear model fit to data.

This function uses lm to fit a linear model to data and outputs the model object. It requires a data table, one quantitative dependent variable and one or more independent variables. The model output can be used to extract coefficients and other information, including post-hoc comparisons. If your experiment design has random factors, use the related function mixed_model.

simple_model(data, Y_value, Fixed_Factor, ...)

Arguments

data
a data table object, e.g. data.frame or tibble.

Y_value
name of column containing quantitative (dependent) variable, provided within "quotes".

Fixed_Factor
name(s) of categorical fixed factors (independent variables) provided as a vector if more than one or within "quotes".

...
any additional arguments to pass on to lm if required.

Value

This function returns the output of lm().

Details

This function is related to link{simple_anova}. Output of this function can be used with posthoc_Pairwise, posthoc_Levelwise and posthoc_vsRef, or with emmeans.

Examples

#Basic usage where the table data_cholesterol is passed with names of one variable within quotes

simple_model(data_cholesterol, "Cholesterol", "Treatment")

#> 
#> Call:
#> lm(formula = Cholesterol ~ Treatment, data = data)
#> 
#> Coefficients:
#>          (Intercept)  TreatmentBefore_drug  
#>               143.93                 12.55  
#> 

#two way ANOVA with Treatment & Hospital as fixed factors
simple_model(data_cholesterol, "Cholesterol", c("Treatment", "Hospital"))

#> 
#> Call:
#> lm(formula = Cholesterol ~ Treatment * Hospital, data = data)
#> 
#> Coefficients:
#>                         (Intercept)                 TreatmentBefore_drug  
#>                             147.011                               26.165  
#>                      HospitalHosp_b                       HospitalHosp_c  
#>                              -4.587                              -42.004  
#>                      HospitalHosp_d                       HospitalHosp_e  
#>                              15.770                               15.432  
#> TreatmentBefore_drug:HospitalHosp_b  TreatmentBefore_drug:HospitalHosp_c  
#>                              -8.075                               -7.001  
#> TreatmentBefore_drug:HospitalHosp_d  TreatmentBefore_drug:HospitalHosp_e  
#>                             -27.202                              -25.803  
#> 

#save model
model <- simple_model(data_cholesterol, "Cholesterol", c("Treatment", "Hospital"))

#get summary
summary(model)

#> 
#> Call:
#> lm(formula = Cholesterol ~ Treatment * Hospital, data = data)
#> 
#> Residuals:
#>     Min      1Q  Median      3Q     Max 
#> -94.992 -25.853  -8.857  20.918 128.431 
#> 
#> Coefficients:
#>                                     Estimate Std. Error t value Pr(>|t|)    
#> (Intercept)                          147.011     23.329   6.302 1.78e-07 ***
#> TreatmentBefore_drug                  26.165     32.992   0.793    0.432    
#> HospitalHosp_b                        -4.587     32.992  -0.139    0.890    
#> HospitalHosp_c                       -42.004     32.992  -1.273    0.210    
#> HospitalHosp_d                        15.770     32.992   0.478    0.635    
#> HospitalHosp_e                        15.432     32.992   0.468    0.642    
#> TreatmentBefore_drug:HospitalHosp_b   -8.075     46.658  -0.173    0.863    
#> TreatmentBefore_drug:HospitalHosp_c   -7.001     46.658  -0.150    0.881    
#> TreatmentBefore_drug:HospitalHosp_d  -27.202     46.658  -0.583    0.563    
#> TreatmentBefore_drug:HospitalHosp_e  -25.803     46.658  -0.553    0.583    
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> 
#> Residual standard error: 52.17 on 40 degrees of freedom
#> Multiple R-squared:  0.156,	Adjusted R-squared:  -0.03385 
#> F-statistic: 0.8218 on 9 and 40 DF,  p-value: 0.5998
#>