In vitro experiments measuring percentage cell death in three genotypes of cells.

These data are from in vitro measurements of death of host cells (measured as percentage of total cells) after infection with three different strains of a pathogenic bacterium, from five independent experiments. The three strains are three levels within the fixed factor Genotype. The five independent experiments are levels within the random variable Experiment. These data can be analysed using linear mixed effects modeling. These data are from Goddard et al, Cell Rep, 2019, doi.org/10.1016/j.celrep.2019.03.100

data_1w_death

Format

data.frame: 15 obs. of 3 variables.

Experiment

Experiment - a random factor with 5 levels "Exp_1","Exp_2"...

Genotype

Genotypes - a fixed factor with 3 levels: "WT","KO_1","KO_2".

Death

Numerical dependent variable indicating percentage cell death.