ANOVA table from linear mixed effects analysis.

This function uses lmer to fit a linear mixed effect model, passes on the model to anova and provides the ANOVA table. It produces a type II sum of squares ANOVA table with Kenward-Roger approximation for degrees of freedom (as implemented in lmerTest) package. It requires a data table, one dependent variable (Y_value), one or more independent variables (Fixed_Factor), and at least one random factor (Random_Factor). This function is related to mixed_model.

mixed_anova(data, Y_value, Fixed_Factor, Random_Factor, ...)

Arguments

data
a data table object, e.g. data.frame or tibble.

Y_value
name of column containing quantitative (dependent) variable, provided within "quotes".

Fixed_Factor
name(s) of categorical fixed factors (independent variables) provided as a vector if more than one or within "quotes".

Random_Factor
name(s) of random factors to allow random intercepts; to be provided as a vector when more than one or within "quotes".

...
any additional arguments to pass on to lmer if required.

Value

This function returns the output of anova.

Details

When more than one fixed factors are entered in the argument, a full model with interaction term is fitted. This means when Y_value = Y, Fixed_factor = c("A", "B"), Random_factor = "R" are entered as arguments, these are passed on as Y ~ A*B + (1|R) (which is equivalent to Y ~ A + B + A:B + (1|R)). For simplicity, only random intercepts are fitted ((1|R)). For factorial ANOVAs the default sum of squares is Type II and degrees of freedom are calculated using the Kenward-Roger approximation. All other settings are lmer and anova defaults.

Examples

#Basic usage where the table data_cholesterol is passed with names of two fixed factors as a vector
mixed_anova(data_cholesterol, "Cholesterol", c("Treatment", "Hospital"), "Subject")

#> Type II Analysis of Variance Table with Kenward-Roger's method
#>                     Sum Sq Mean Sq NumDF DenDF F value    Pr(>F)    
#> Treatment          1968.38 1968.38     1    20 46.2485 1.301e-06 ***
#> Hospital            131.43   32.86     4    20  0.7720 0.5561366    
#> Treatment:Hospital 1482.35  370.59     4    20  8.7072 0.0003066 ***
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1

#Usages with one fixed (Student) and random factor (Experiment), each within quotes
mixed_anova(data_doubling_time, "Doubling_time", "Student", "Experiment")

#> boundary (singular) fit: see ?isSingular

#> Type II Analysis of Variance Table with Kenward-Roger's method
#>         Sum Sq Mean Sq NumDF DenDF F value Pr(>F)
#> Student 16.824  1.8694     9    18  0.4725 0.8744