Model from a linear mixed effects model with varying slopes

One of four related functions for mixed effects analyses (based on lmer and as_lmerModLmerTest) to get a linear model for downstream steps, or an ANOVA table.

mixed_model

mixed_anova

mixed_model_slopes

mixed_anova_slopes.

mixed_model_slopes(
  data,
  Y_value,
  Fixed_Factor,
  Slopes_Factor,
  Random_Factor,
  AvgRF = TRUE,
  ...
)

Arguments

data

a data table object, e.g. data.frame or tibble.

Y_value

name of column containing quantitative (dependent) variable, provided within "quotes". The following transformations are permitted: "log(Y_value)", "log(Y_value + c)" where c a positive number, "logit(Y_value)" or "logit(Y_value/100)" which may be useful when Y_value are percentages (note quotes outside the log or logit calls); "sqrt(Y_value)" or "(Y_value)^2" should also work. During posthoc-comparisons, log and logit transformations will be back-transformed to the original scale. Other transformations, e.g., "sqrt(Y_value)" will not be back-transformed. Check out the regrid and ref_grid for details if you need back-transformation to the response scale.

Fixed_Factor

name(s) of categorical fixed factors (independent variables) provided within quotes (e.g., "A") or as a vector if more than one (e.g., c("A", "B"). If a numeric variable is used, transformations similar to Y_value are permitted.

Slopes_Factor

name of factor to allow varying slopes on. One one variable is allowed.

Random_Factor

name(s) of random factors to allow random intercepts; to be provided within quotes (e.g., "R") or as a vector when more than one (e.g., c("R1", "R2")). Only one variable is allowed.

AvgRF

this is a new argument since v5.0.0. The default AvgRF = TRUE will use the mean of Y_value (the response variable) grouped by levels of the Fixed_Factor and Random_Factor (using table_summary). This ensures that replicates within Random_Factor (or any other unused variable) are averaged (e.g., technical replicates nested within experimental blocks) before fitting a linear model and the denominator Df values are sensible. The name of the data frame in the model object will have (AvgRF) appended to it to indicate the averaging within levels of the Random_Factor. Using AvgRF = FALSE will lead to behaviour like versions <5.0.0.

...

any additional arguments to pass on to lmer if required.

Value

This function returns an S4 object of class "lmerModLmerTest".

Details

These functions require a data table, one dependent variable (Y_value), one or more independent variables (Fixed_Factor), and at least one random factor (Random_Factor). These should match names of variables in the long-format data table exactly. Since v5.0.0, if AvgRF = TRUE, the response variable is averaged over levels of the fixed and random factors (to collapse replicate observations) and reduce the number of denominator degrees of freedom. If you do not want to do this, set AvgRF = FALSE.

For more advanced models with slopes and intercept, use mixed_model or mixed_anova using the Formula argument.

Outputs of mixed_model and mixed_model_slopes can be used for post-hoc comparisons with posthoc_Pairwise, posthoc_Levelwise, posthoc_vsRef, posthoc_Trends_Pairwise, posthoc_Trends_Levelwise and posthoc_Trends_vsRefor with emmeans.

More than one fixed factors can be provided as a vector (e.g. c("A", "B")). A full model with interaction term is fitted. This means when Y_value = Y, Fixed_factor = c("A", "B"), Random_factor = "R" are entered as arguments, these are passed on as Y ~ A*B + (1|R) (which is equivalent to Y ~ A + B + A:B + (1|R)).

In mixed_model_slopes and mixed_anova_slopes, the following kind of formula is used: Y ~ A*B + (S|R) (which is equivalent to Y ~ A + B + A:B + (S|R)). In this experimental implementation, random slopes and intercepts are fitted ((Slopes_Factor|Random_Factor)). Only one term each is allowed for Slopes_Factor and Random_Factor.

Examples

#two fixed factors as a vector, 
#exactly one slope factor and random factor
mod <- mixed_model_slopes(data = data_2w_Tdeath,
Y_value = "PI",
Fixed_Factor = c("Genotype", "Time"),
Slopes_Factor = "Time",
Random_Factor = "Experiment")
#> The new argument `AvgRF` is set to TRUE by default in >=5.0.0). See help for details.
#> boundary (singular) fit: see help('isSingular')
#get summary
summary(mod)
#> Linear mixed model fit by REML. t-tests use Satterthwaite's method [
#> lmerModLmerTest]
#> Formula: PI ~ Genotype * Time + (Time | Experiment)
#>    Data: data_2w_Tdeath (AvgRF)
#> 
#> REML criterion at convergence: 132.9
#> 
#> Scaled residuals: 
#>      Min       1Q   Median       3Q      Max 
#> -1.74243 -0.50144  0.01573  0.55592  1.88600 
#> 
#> Random effects:
#>  Groups     Name        Variance Std.Dev. Corr
#>  Experiment (Intercept)  7.372   2.715        
#>             Timet300     1.231   1.110    1.00
#>  Residual               24.327   4.932        
#> Number of obs: 24, groups:  Experiment, 6
#> 
#> Fixed effects:
#>                     Estimate Std. Error      df t value Pr(>|t|)    
#> (Intercept)           23.841      2.299  12.624  10.373 1.53e-07 ***
#> GenotypeKO           -16.918      2.848  15.000  -5.941 2.71e-05 ***
#> Timet300              13.563      2.883  14.912   4.704 0.000287 ***
#> GenotypeKO:Timet300  -10.293      4.027  15.000  -2.556 0.021937 *  
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> 
#> Correlation of Fixed Effects:
#>             (Intr) GntyKO Tmt300
#> GenotypeKO  -0.619              
#> Timet300    -0.536  0.494       
#> GntyKO:T300  0.438 -0.707 -0.698
#> optimizer (nloptwrap) convergence code: 0 (OK)
#> boundary (singular) fit: see help('isSingular')
#>