Plot a dotplot on a violin plot with two variables.

This function takes a data table, X and Y variables, and plots a graph with a dotplot and violinplot using ggplot.

plot_dotviolin(
  data,
  xcol,
  ycol,
  dotsize = 1,
  trim = T,
  scale = "area",
  fontsize = 20,
  alpha = 1,
  ColPal = "all_grafify",
  ColRev = F
)

Arguments

data
a data table object, e.g. data.frame or tibble

xcol
name of the column to plot on X axis. This should be a categorical variable.

ycol
name of the column to plot on quantitative Y axis. This should be a quantitative variable.

dotsize
size of dots relative to binwidth used by geom_dotplot. Default set to 1, increase/decrease as needed.

trim
set whether tips of violin plot should be trimmed at high/low data. Default trim = T, can be changed to F.

scale
set to "area" by default, can be changed to "count" or "width".

fontsize
parameter of base_size of fonts in theme_classic, default set to size 20.

alpha
fractional opacity of violins, default set to 1 (i.e. maximum opacity & zero transparency)

ColPal
grafify colour palette to apply, default "all_grafify"; alternatives: "okabe_ito", "bright", "pale", "vibrant", "contrast", "muted" "dark", "light".

ColRev
whether to reverse order of colour choice, default F (FALSE); can be set to T (TRUE)

Value

This function returns a ggplot2 object on which additional geometries etc. can be added.

Details

The function uses geom_violin and geom_dotplot geometries. Note that the geom_violin options are set as follows: scale = "area", draw_quantiles = c(0.25, .5, .75). The trim = T set by default can be changed when calling the function. The X variable is mapped to the fill aesthetic in both violinplot and dotplot, and its colour can be changed using scale_fill_brewer or any scale_fill... option. The size of dots can be adjusted using the parameter, which is dotsize = 1 by default.

This function is related to plot_scatterbar_sd, plot_dotbar_sd and plot_dotviolin.

Examples

#Basic usage requires a data table and X & Y variables#'
#plot_dotviolin(data_doubling_time, Student, Doubling_time)

# or with trim = F
#plot_dotviolin(data_doubling_time, Student, Doubling_time, trim = F)