Plot model diagnostics for generalised additive models

This is a clone of the appraise function in the gratia package (rewritten to avoid depending on gratia package for these plots). This function will plot 4 diagnostic plots when given a generalised additive model fitted with ga_model or mgcv. It creates graphs that use grafify colours and theme_grafify().

plot_qq_gam(
  Model,
  symsize = 2,
  s_colour = "#E69F00",
  s_alpha = 0.6,
  line_col = "black",
  base_size = 12,
  linethick,
  n_bins = c("sturges", "scott", "fd")
)

Arguments

Model

a model of class gam fitted with ga_model or the mgcv package.

symsize

size of symbols (default = 2)

s_colour

colour of symbols (default = ok_orange)

s_alpha

opacity of symbols (default = 0.8)

line_col

colour of lines (default = black)

base_size

font size for theme (default = 12)

linethick

thickness in 'pt' units of lines and symbol orders (default = base_size/22)

n_bins

one of either "sturges", "scott", "fd"

Value

This function returns an object of classes "ggplot" and "gg".

This function returns a ggplot2 object of class "gg" and "ggplot".