Level-wise post-hoc comparisons from a linear or linear mixed effects model.

This function is a wrapper based on emmeans, and needs a model produced by simple_model (which calls lm) or a linear mixed effects model produced by mixed_model (which calls lmer). It also needs to know the fixed factors, which should match those in the model.

posthoc_Levelwise(Model, Factors, P_Adj = "fdr", ...)

Arguments

Model
a model object fit using simple_model or mixed_model (or lm or lmer).

Factors
one or more categorical variables, provided as a vector (see Examples), whose levels you wish to compare pairwise. Names of factors should match factors used to fit the model. When more than one factor is provided e.g. Fixed_factor = c("A", "B"), this function passes this on as specs = A|B (note the vertical | between the two factors) to emmeans to produce comparisons between each level A with each other listed separately at each level of B.

P_Adj
method for correcting P values for multiple comparisons. Default is set to false discovery rate ("fdr"), can be changed to "none", "tukey", "bonferroni", "sidak". See the manual for emmeans.

...
additional arguments for emmeans such as lmer.df or others. See help for sophisticated models in emmeans.

Value

returns results produced by emmeans.

Details

The function will generate level-wise comparisons, i.e. comparison between of every level of one factor separately at each level of the other factor. By default, P values are corrected by the FDR method (which can be changed). If the model was fit by transforming the quantitative response variable using "log", "logit", "sqrt" etc., results will still be on the original scale, i.e. type = "response" is the default; data will be back-transformed (check results to confirm this), and for log or logit, ratios will be compared. The first part of the emmeans results has the estimated marginal means, SE and CI ($emmeans), which are generated from the fitted model, and not the original data table. The second part has the results of the comparisons ($contrasts).

Examples

#basic use with one factor
#note quotes used only for fixed factors
#to get comparisons between different hospitals separately for each level of Treatment
#posthoc_Levelwise(CholMod, c("Hospital", "Treatment"))

#get comparisons between treatments separately at each hospital
#posthoc_Levelwise(CholMod, c("Treatment", "Hospital"))