Post-hoc comparisons to a control or reference group.

This function is a wrapper based on emmeans, and needs a model produced by simple_model (which calls lm) or a linear mixed effects model produced by mixed_model (which calls lmer). It also needs to know the fixed factors, which should match those in the model.

posthoc_vsRef(Model, Factors, Ref = 1, P_Adj = "fdr", ...)

Arguments

Model
a model object fit using simple_model or mixed_model (or lm or lmer).

Factors
Factors one or more categorical variables, provided as a vector (see Examples), whose levels you wish to compare pairwise. Names of factors should match factors used to fit the model. When more than one factor is provided e.g. Fixed_factor = c("A", "B"), this function passes this on as specs = A|B (note the vertical | between the two factors) to emmeans. The specification internally is set to specs = trt.vs.ctrl, ref = 1 to compare each group in A to the reference first group in A, separately at each level of B.

Ref
the reference group provided as a number; default ref = 1.

P_Adj
P_Adj method for correcting P values for multiple comparisons. Default is set to false discovery rate ("fdr"), can be changed to "none", "tukey", "bonferroni", "sidak". See the manual for emmeans

...
additional arguments for emmeans such as lmer.df or others. See help for sophisticated models in emmeans.

Value

returns the result of emmeans contrasts.

Details

The function will generate treatment vs control type of comparisons, i.e. comparison of each level of a factor to a reference level, which is set by default to the first level in the factor (ref = 1). By default, P values are corrected by the FDR method (which can be changed). If the model was fit by transforming the quantitative response variable using "log", "logit", "sqrt" etc., results will still be on the original scale, i.e. type = "response" is the default. Data will be back-transformed (check results to confirm this), and for log or logit, ratios will be compared. The first part of the emmeans results has the estimated marginal means, SE and CI ($emmeans), which are generated from the fitted model, and not the original data table. The second part has the results of the comparisons ($contrasts).

Examples

#basic usage to compare against ref = 1
#posthoc_vsRef(DoublMod, "Student")

#to compare all students with student #9
#posthoc_vsRef(DoublMod, "Student", ref = 9)

#for comparison between hospital #1 to every other hospital, separately at levels of Treatment
#posthoc_vsRef(CholMod, c("Hospital", "Treatment"))

#for comparisons between reference group of treatment to every other group, 
#listed separated at every level of Hospital
#posthoc_vsRef(CholMod, c("Treatment", "Hospital"))