ANOVA table from a linear model fit to data.

Update in v0.2.1: This function uses lm to fit a linear model to data, passes it on to Anova, and outputs the ANOVA table with type II sum of squares with F statistics and P values. (Previous versions produced type I sum of squares using anova call.)

simple_anova(data, Y_value, Fixed_Factor, ...)

Arguments

data
a data table object, e.g. data.frame or tibble.

Y_value
name of column containing quantitative (dependent) variable, provided within "quotes".

Fixed_Factor
name(s) of categorical fixed factors (independent variables) provided as a vector if more than one or within "quotes".

...
any additional argument to ass on to lm if required.

Value

ANOVA table output by anova.

Details

It requires a data table, one quantitative dependent variable and one or more independent variables. If your experiment design has random factors, use the related function mixed_anova.

This function is related to link{simple_model}.

Examples

#Basic usage where the table data_doubling_time is passed with names of the fixed factor within quotes

simple_anova(data_doubling_time, "Doubling_time", "Student")

#> Anova Table (Type II tests)
#> 
#> Response: Doubling_time
#>           Sum Sq Df F value Pr(>F)
#> Student   16.824  9  0.4725 0.8761
#> Residuals 79.126 20